+++ src/database/operations/load/index.ts
@@ -6,6 +6,8 @@ export * from './loadBaseFields';
 export * from './loadBulkUpload';
 export * from './loadBulkUploadBundle';
 export * from './loadBundle';
+export * from './loadOpportunity';
+export * from './loadOpportunityBundle';
 export * from './loadOrganization';
 export * from './loadOrganizationBundle';
 export * from './loadOrganizationByEmployerIdentificationNumber';