+++ src/database/queries/opportunities/insertOne.sql
@@ -1,4 +1,2 @@
 INSERT INTO opportunities ( title ) VALUES ( :title )
-RETURNING id AS "id",
-  title as "title",
-  created_at as "createdAt";
+RETURNING opportunity_to_json(opportunities) AS "object";