+++ src/database/queries/opportunities/selectById.sql
@@ -1,5 +1,3 @@
-SELECT o.id AS "id",
-  o.title AS "title",
-  o.created_at AS "createdAt"
-FROM opportunities o
-WHERE o.id = :id;
+SELECT opportunity_to_json(opportunities) AS "object"
+FROM opportunities
+WHERE id = :id;