+++ src/types/Opportunity.ts
@@ -1,5 +1,29 @@
-export interface Opportunity {
-	id: number;
+import { ajv } from '../ajv';
+import type { JSONSchemaType } from 'ajv';
+import type { Writable } from './Writable';
+
+interface Opportunity {
+	readonly id: number;
 	title: string;
-	createdAt: Date;
+	readonly createdAt: string;
 }
+
+type WritableOpportunity = Writable<Opportunity>;
+
+const writableOpportunitySchema: JSONSchemaType<WritableOpportunity> = {
+	type: 'object',
+	properties: {
+		title: {
+			type: 'string',
+		},
+	},
+	required: ['title'],
+};
+const isWritableOpportunity = ajv.compile(writableOpportunitySchema);
+
+export {
+	Opportunity,
+	WritableOpportunity,
+	writableOpportunitySchema,
+	isWritableOpportunity,
+};